Update Rickshaw.Graph.js
Fixing an apparent bug in this.stackData: when we are toggling between series of different scales the data[index] access at line 149 can hit an array out-of-bounds error if there are inactive series.  Added active() to the call at line 146, to parallel the call at line 130.  This seems to fix the issue.