Constrains hover detail and label to the graph area
In cases where the xLabel or item elements would extend past the right
edge of the graph element, they are aligned with their right edges on
the indicator line, rather than their left edges.  This should produce
acceptable behavior in any case where those elements are able to fit
inside the graph area.  If one or both of the elements would not fit,
the behavior will be no worse than the previous behavior.