Rescope stroke style attribute on y-axis
In mbostock/d3 issue #1016, the scope of the `tick` class on a svg y-axis
element changed to include the entire g object (line and text), not just the
line. The Rickshaw CSS for `.y_ticks .tick` was created in the early commits, and
sets the style to be a 2px width semitransparent stroke.

Due to the change in scoping of the style class, the stroke element applies to
both the line and the text element, creating a shadow on the y-axis text labels.

By changing the scope of the stroke stanza from `.y_ticks .tick` to `.y_ticks
.tick line`, the stroke now applies only to the line, and not to the text
element, this removing the stroke shadow on the text label.